Only include the AWS SDK Core packages (#12)
opentracing-aws-sdk included the entire AWS SDK distribution as
dependencies. This seems unnacessary, as the TracingRequestHandler
only really depends on sdk-core. Therefore, include only sdk-core
as a direct dependency.

The dynamodb SDK is included as a test dependency so unittests pass.

Also upgrade versions of SDK 1 and SDK 2 to the current versions.